ArguEdit

Description

The ArguEdit module is the actual programm controler, which binds all other modules. The Controler gets to react on user input coming from the ArgView module and passing it on to the parser following the Model. Thus it can be regarded as Model View Controler architecture implementation in a loose sense of the term.

ArguEdit loads a glade file for the basic visualization which is added the different GTK representations of argument components like thesis, premise or definition.

The main container allows for loading from and saving to a file as well as other clickings. Hover the status icons to reveal operations on the components. Clicking the status icons changes the status icon.

Documentation

start :: IO ()