ArguView

Description

This is the module that contains the major viewing functionality for Arguedit. That is, the argumentation components need a visual representation such as text boxes, clickable status icons and tree structured contains relation to have components belong to each other as wished.

Documentation

data GtkParts

A graphical element of the argumentation consists of gtk parts such as boxes, textviews, entry fields adn so on. This structures binds them together deviding into two categories. The common parts are those that all elements have in common where GtkDefSpecific is specific to the definition element.

Constructors

GtkCommonParts

Fields

ioRefHimg :: IORef HeaderImage

ioRefMeta :: IORef Meta

parent :: VBox

row :: HBox

tbox :: VBox

tview :: TextView

GtkDefSpecific

Fields

term :: Entry

data HeaderImage

The header's image is a structure to combine information about the element button.

Constructors

HeaderImage

Fields

imgLetter :: ImgLetter

imgColors :: [ImgColor]

imgSize :: Int

updateMeta :: GtkParts -> Meta -> IO ()

updateMeta is function to be run in the IO monade updating the meta information of an element which is placed in a static variable.

getMeta :: GtkParts -> IO Meta

Get the meta value from the static variable stored in the GtkParts object.

getHimg :: GtkParts -> IO HeaderImage

Get the header's image from a GtkParts' static variable.

nextImage :: GtkParts -> IO Image

An OnClick event triggers this function to change the current image to the next one available.

fileNameFromHeaderImage :: HeaderImage -> [Char]

Builds a filepath and filename for given HeaderImage by its letter and color.

toLower :: [Char] -> [Char]

Convenience function to lower the case of a string.

data ArgHeaderView

ArgHeaderView is the graphical representation of element in the argumentation.

Constructors

EmptyView

ArgHeaderThesisView

Fields

commonParts :: GtkParts

ArgHeaderCntrExmplView

Fields

commonParts :: GtkParts

ArgHeaderLogicView

Fields

commonParts :: GtkParts

ArgHeaderPremiseView

Fields

commonParts :: GtkParts

ArgHeaderPredicateView

Fields

commonParts :: GtkParts

ArgHeaderRationaleView

Fields

commonParts :: GtkParts

ArgHeaderDefView

Fields

commonParts :: GtkParts

specificParts :: GtkParts

getText :: TextViewClass self => self -> IO String

setText :: TextViewClass self => self -> String -> IO ()

newTextView :: String -> IO TextView

newTextBuffer :: String -> IO TextBuffer

textBufferSetFontSize :: TextBufferClass self => self -> Double -> IO ()

newImage :: String -> Int -> IO Image

newRow :: IO HBox

The elements are organized in rows and columns. This fct. creates a new row.

newCol :: IO VBox

The elements are organized in rows and columns. This fct. creates a new column.

getStockIcon :: [Char] -> Maybe StockId

Some images belong to the systems stock. Basically the strings New, Open, Save, ... are mapped to stock plus string here which are system known values.

addToBox :: (BoxClass box, WidgetClass child) => box -> child -> Packing -> IO ()

Finally gtk parts must be added to a container such as box. This function adds a child to be contained.

newTextBox :: t -> String -> IO (VBox, VBox, TextView)

Create a new textbox represented as a triplet of outer column, innerbox and textview element for more convenient access.

newDefBox :: String -> String -> IO (VBox, Entry, TextView)

Create a definition box with term to be defined and its defining description.

install_main_menu :: BoxClass self => (t, self, t1, ActionGroup) -> IO ()

prAct :: ActionClass self => self -> IO (ConnectId self)

createMainGUI :: (Window -> Box -> VBox -> IO ActionGroup) -> IO ()

Create the main gui from a glade file and installing a menu to start with.

toggleShowHideView :: ArgHeaderView -> IO ()

Toggle the visibility of the element's row

toggleVisibility :: WidgetClass a => a -> IO ()

Wrapper for the builtin show and hide functionality.

viewMain :: (Window -> Box -> VBox -> IO ActionGroup) -> IO ()

Begin viewing everything by initilizing the GTK components of the gu| Begin viewing the everything by initilizing the GTK components of the gui.

viewIconSizeThesis :: Int

viewIconSizeDefinition :: Int

viewIconSizeDefault :: Int

viewIconSizeMini :: Int

uiDecl :: [Char]

Constant string to hold the menu structure. For convenience reasons not in a separated file.