Gtk.MenuCreator

Description

The MenuCreator module is a mere helper module which provides menu creation as it grows out of the view code to be reused on different pages in consideration.

Documentation

createMenuBar :: [([Char], [([Char], Maybe (IO ()))])] -> IO MenuBar

default_menuBarDescr :: [([Char], [([Char], Maybe (IO ()))])]