Indentparser.ArguParser

Description

This is the Parser for the argumentations ArguEdit gets to work with. It also specifies the fileformat of the argumentations files *.arg .

It recognizes indentation as part of the content as definitions premises and so on are contained inside the discourse on a thesis.

The file format looks like this: ArgComp identifier(optional Status status value): String (optional indented lines as list of items)

(Note: On load the lack of an explicit Status implies Status Initial.)

Documentation

parseArgumentation :: FilePath -> IO (Either ParseError [ArgComp])

parseSpeechFile :: FilePath -> IO (Either ParseError [Speech])

parseDialogueFile :: FilePath -> IO (Either ParseError [Dialogue])