Specification.ArguModel

Description

This the core model of ArguEdit. Basic structure of the transcription of argumentations as well as fundamental operations on them can be found here.

Documentation

data ImgLetter

Constructors

A

B

C

D

E

F

G

H

I

J

K

L

M

N

O

P

Q

R

S

T

U

V

W

X

Y

Z

data ImgColor

Constructors

Black

Blue

DarkGreen

LightGreen

Gold

Grey

Orange

Pink

Red

Violet

availableColors :: [ImgColor]

statColorMap :: Bimap ImgColor EStatus

whichColor :: Monad m => EStatus -> m ImgColor

whichEStatus :: Monad m => ImgColor -> m EStatus

updateListAttribute :: Attribute -> [Attribute] -> [Attribute]

replaceAttr :: Attribute -> Attribute -> Attribute

conformHead :: Attribute -> [ImgColor]

conformHeadColor :: ImgColor -> [ImgColor] -> [ImgColor]

data EStatus

Constructors

Initial

Valid

Invalid

Verified

Proposed

Falsified

data Attribute

Constructors

Label String

Author String

ID Int

Status EStatus

statFromMeta :: Meta -> Attribute

data Meta

Constructors

NoMeta

Meta [Attribute]

data Text

Constructors

Text String

data Descr

Constructors

Descr String

data Term

Constructors

Term String

data ArgContent

Constructors

Thesis Text

CounterExample Text

Logic Text

Premise Text

Predicate Text

Rationale Text

Definition Term Descr

availableLettersDE :: [ImgLetter]

availableLettersEN :: [ImgLetter]

data Name

Constructors

Name String

data SpeakerID

Constructors

SPID Int

data Speaker

Constructors

Speaker Name

data SpeakerList

Constructors

SpeakerList [(SpeakerID, Speaker)]

data Speech

Constructors

Speech Attribute [Text]

data Dialogue

Constructors

Dialogue Attribute [Speech]

data Header

Constructors

ArgHeader Meta ArgContent

DlgHeader Dialogue

data MoveDirection

Constructors

Down

Up

type ArgCompForest = [ArgComp]

data ArgComp

Constructors

ArgComp Header ArgCompForest

type ArgElemId = Int

type ArgPairForest a b = [Tree (a, b)]

type ArgPairTree a b = Tree (a, b)

test_id_tree :: Tree (String, ArgElemId)

removeById :: Eq b => [Tree (a, b)] -> b -> [Tree (a, b)]

moveById :: Eq b => Tree (a, b) -> b -> MoveDirection -> Tree (a, b)

moveChildInListById :: Eq a => [Tree (t, a)] -> a -> MoveDirection -> [Tree (t, a)]

moveInListById :: Eq b => [Tree (a, b)] -> b -> MoveDirection -> [Tree (a, b)]

isIn :: Eq b => b -> [Tree (a, b)] -> Bool

compareNodeId :: Eq a => Tree (t, a) -> a -> Bool

prependById :: Eq b => [Tree (a, b)] -> Tree (a, b) -> b -> [Tree (a, b)]

appendById :: Eq b => [Tree (a, b)] -> Tree (a, b) -> b -> [Tree (a, b)]

addById :: Eq b => [Tree (a, b)] -> Tree (a, b) -> b -> [Tree (a, b)]

getNodeById :: Eq b => Tree (a, b) -> b -> Maybe (Tree (a, b))

getNodeByIdFromList :: Eq b => [Tree (a, b)] -> b -> Maybe (Tree (a, b))

getChildrenById :: Eq b => [Tree (a, b)] -> b -> [Tree (a, b)]

adoptChild :: ArgPairTree a b -> ArgPairTree a b -> ArgPairTree a b

type OpCount = Int

type ArgCompState = ([ArgComp], OpCount)

initialState :: ArgCompState

add :: ArgComp -> State ArgCompState OpCount

replaceACList :: [ArgComp] -> State ArgCompState OpCount

drawArgComp :: ArgComp -> [String]

drawArgCompForest :: [ArgComp] -> IO ()

drawTreeOne :: ArgComp -> IO ()

drawTreeOneToFile :: ArgComp -> Handle -> IO ()

drawTreeToFile :: [ArgComp] -> Handle -> IO ()

writeACToFile :: FilePath -> [ArgComp] -> IO ()

mt1 :: Meta

test_dialogue :: Header