ArguView

Description

This is the module that contains the major viewing functionality for Arguedit. That is, the argumentation components need a visual representation such as text boxes, clickable status icons and tree structured contains relation to have components belong to each other as wished.

Documentation

data GtkParts

Constructors

GtkCommonParts

Fields

ioRefHimg :: IORef HeaderImage

ioRefMeta :: IORef Meta

parent :: VBox

row :: HBox

tbox :: VBox

tview :: TextView

GtkDefSpecific

Fields

term :: Entry

data HeaderImage

Constructors

HeaderImage

Fields

imgLetter :: ImgLetter

imgColors :: [ImgColor]

imgSize :: Int

updateMeta :: GtkParts -> Meta -> IO ()

getMeta :: GtkParts -> IO Meta

getHimg :: GtkParts -> IO HeaderImage

nextImage :: GtkParts -> IO Image

fileNameFromHeaderImage :: HeaderImage -> [Char]

toLower :: [Char] -> [Char]

data ArgHeaderView

Constructors

EmptyView

ArgHeaderThesisView

Fields

commonParts :: GtkParts

ArgHeaderCntrExmplView

Fields

commonParts :: GtkParts

ArgHeaderLogicView

Fields

commonParts :: GtkParts

ArgHeaderPremiseView

Fields

commonParts :: GtkParts

ArgHeaderPredicateView

Fields

commonParts :: GtkParts

ArgHeaderRationaleView

Fields

commonParts :: GtkParts

ArgHeaderDefView

Fields

commonParts :: GtkParts

specificParts :: GtkParts

getText :: TextViewClass self => self -> IO String

setText :: TextViewClass self => self -> String -> IO ()

newTextView :: String -> IO TextView

newTextBuffer :: String -> IO TextBuffer

textBufferSetFontSize :: TextBufferClass self => self -> Double -> IO ()

newImage :: String -> Int -> IO Image

newRow :: IO HBox

newCol :: IO VBox

getStockIcon :: [Char] -> Maybe StockId

addToBox :: (BoxClass box, WidgetClass child) => box -> child -> Packing -> IO ()

newTextBox :: t -> String -> IO (VBox, VBox, TextView)

newDefBox :: String -> String -> IO (VBox, Entry, TextView)

install_main_menu :: BoxClass self => (t, self, t1, ActionGroup) -> IO ()

prAct :: ActionClass self => self -> IO (ConnectId self)

createMainGUI :: (Window -> Box -> VBox -> IO ActionGroup) -> IO ()

toggleShowHideView :: ArgHeaderView -> IO ()

toggleVisibility :: WidgetClass a => a -> IO ()

viewMain :: (Window -> Box -> VBox -> IO ActionGroup) -> IO ()

viewIconSizeThesis :: Int

viewIconSizeDefinition :: Int

viewIconSizeDefault :: Int

viewIconSizeMini :: Int

uiDecl :: [Char]